implement common dataframe column api class
